% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_results.R
\name{create_risk_df}
\alias{create_risk_df}
\title{Create a Data Frame of Operating Characteristic Risks}
\usage{
create_risk_df(oc_all, lrv, tv, group_vars = NULL)
}
\arguments{
\item{oc_all}{A data frame from \code{process_sim_results()}.}

\item{lrv}{The Lower Reference Value used in the simulation.}

\item{tv}{The Target Value used in the simulation.}

\item{group_vars}{A character vector of variable names to group by. If \code{NULL} (default),
the function will automatically use all setting columns.}
}
\value{
A data frame with the FGR and FSR for each simulation scenario.
}
\description{
This function takes the operating characteristics from a simulation run and
creates a data frame summarizing the False Go Rate (FGR) and False
Stop Rate (FSR) for each simulation scenario.
}
