% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{data_gen_binary}
\alias{data_gen_binary}
\title{Generate Binary Response Data}
\usage{
data_gen_binary(
  n_arms = 2,
  nsim = 10,
  n_list = NULL,
  prob_list = NULL,
  arm_names = NULL
)
}
\arguments{
\item{n_arms}{A scalar. The number of arms to be generated in the simulation.}

\item{nsim}{A scalar. The number of repetitions of the simulation to be performed.}

\item{n_list}{A list. Each element represents the number of patients in each arm.}

\item{prob_list}{A list. Each element represents the probability of response (binary outcome = 1) for each arm.}

\item{arm_names}{A list. Each element represents the name of the corresponding arm.}
}
\value{
A list of data frames:
  - \code{data}: A data frame containing individual-level binary outcomes (0/1) for all arms and all repetitions of the simulation.
  - \code{true_value}: A data frame of true response rates for each arm, along with the difference in response rates between the treatment and control arms (if both are present).
  - \code{freq}: A data frame of response frequencies, where each row represents the number of responses and the total number of patients in each arm for each simulation repetition.
}
\description{
This function simulates binary endpoint data (0/1) for multiple study arms across several simulation repetitions.
Each arm's data is generated based on the specified probabilities of response, and the function returns
individual-level data as well as summary statistics for the true response rates and response frequencies.
}
\examples{
# Example of generating binary response data for 2 arms, 100 patients per arm,
# and 10 simulation repetitions:
n_list <- c(treatment = 50, control = 50, control_h = 200)
prob_list <- c(treatment = 0.4, control = 0.2, control_h = 0.3)
arm_names <- c(treatment = "treatment", control = "control", control_h = "control_h")
sim_data_bin <- data_gen_binary(n_arms = 3, nsim = 10, n_list = n_list,
                                prob_list = prob_list, arm_names = arm_names)

}
