% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscore_func.R
\name{eval_gscore_approx_dist}
\alias{eval_gscore_approx_dist}
\title{Evaluate g-score Median Ratio Estimation}
\usage{
eval_gscore_approx_dist(sim_data)
}
\arguments{
\item{sim_data}{A data frame. Data simulated by the `data_gen_gscore()` function.}
}
\value{
A data frame containing evaluation metrics for the g-score median ratio:
  - bias_avg: Average bias of the log-median ratio estimates.
  - bias_avg_median_ratio1: Average bias of the median ratio, obtained by exponentiating the average of log median ratio estimates and subtracting the true median ratio
  - bias_avg_median_ratio2: Average bias of the median ratio, obtained by exponentiating the log median estimates for both arms to obtain the median ratio estimates for each repetition and taking the average bias over all repetitions of simulation
  - sd_avg: Average standard error of the log-median ratio estimates.
  - sd_empirical: Empirical standard deviation of the log-median ratio estimates.
  - cp: Coverage probability for the 95% confidence interval.
}
\description{
This function evaluates the reliability of the normal approximation for g-score median estimates.
For each simulation repetition, it computes the log-median ratio and evaluates metrics such as bias, standard error, and coverage probability.
}
\examples{
# Generating sim_data by function data_gen_gscore()
n_list <- c(treatment = 100, control = 100)
prob_list <- c(treatment = 0.2, control = 0.3)
mu_list <- c(treatment = 0.5, control = 0.6)
sigma_list <- c(treatment = 0.1, control = 0.2)
arm_names <- c(treatment = "treatment", control = "control")
sim_data <- data_gen_gscore(n_arms = 2, nsim = 10, n_list = n_list,
                            prob_list = prob_list, mu_list = mu_list,
                            sigma_list = sigma_list, arm_names = arm_names)

eval_gscore_approx_dist(sim_data = sim_data)

}
