% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{freq_binary}
\alias{freq_binary}
\title{Calculate Response Frequencies for Binary Endpoints}
\usage{
freq_binary(data)
}
\arguments{
\item{data}{A data frame. Individual-level binary outcomes (0/1) for all arms and repetitions of the simulation.}
}
\value{
A data frame. Each row represents the number of responses and the total number of patients in each arm for each simulation repetition.
The data frame includes the following columns:
  - `nsim`: Simulation repetition number.
  - `arm.n`: Total number of patients in each arm.
  - `arm.count`: Total number of responses in each arm.
}
\description{
This function calculates the frequency of binary endpoint responses (0/1) across arms for multiple simulation repetitions.
It groups the data by arm and simulation repetition to count the number of responses and the total number of patients in each arm.
}
\examples{
data <- data.frame(nsim = rep(1:10, each = 200),
                   arm = rep(c("A","B"), each = 100, times = 10),
                   resp = stats::rbinom(2000, 1, 0.5))
freq_binary(data = data)

}
