% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_in}
\alias{get_in}
\title{Safely get an element from a (possibly NULL) list}
\usage{
get_in(lst, key)
}
\arguments{
\item{lst}{A list (possibly \code{NULL}).}

\item{key}{Element name to extract.}
}
\value{
The element \code{lst[[key]]} or \code{NULL}.
}
\description{
Returns \code{NULL} if the list itself is \code{NULL}.
}
\examples{
get_in(list(a = 1), "a")
get_in(NULL, "a")
}
