% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscore_func.R
\name{get_true_gscore_median}
\alias{get_true_gscore_median}
\title{Calculate True Median of g-scores}
\usage{
get_true_gscore_median(p, mu, sigma)
}
\arguments{
\item{p}{A scalar. Probability of a g-score being zero.}

\item{mu}{A scalar. Mean value of the normal distribution for generating log of positive g-scores.}

\item{sigma}{A scalar. Standard deviation of the normal distribution for generating log of positive g-scores.}
}
\value{
A scalar. True median of the g-scores.
}
\description{
This function calculates the true median of g-scores generated by a log-normal distribution, accounting for zero inflation.
}
\examples{
get_true_gscore_median(p = 0.3, mu = 1, sigma = 0.5)
}
