% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_oc.R
\name{obtain_oc}
\alias{obtain_oc}
\title{Obtain Operating Characteristics (OC)}
\usage{
obtain_oc(decisions)
}
\arguments{
\item{decisions}{A data frame. Contains the decisions made in each simulation or trial, with columns for posterior probability-based decisions (\code{decision_pr}) and credible interval-based decisions (\code{decision_ci}).}
}
\value{
A data frame that includes the original settings along with the calculated proportions of decisions based on the posterior probability and credible interval. The output includes:
\itemize{
\item \code{proportion_pr}: Proportion of decisions made based on the posterior probability.
\item \code{proportion_ci}: Proportion of decisions made based on the credible interval.
}
}
\description{
This function calculates the operating characteristics (OC) by computing the proportion of decisions made based on the posterior probability and the credible interval from a set of decision results.
It combines these proportions with the provided settings data to return a comprehensive summary.
}
