% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_risk}
\alias{plot_risk}
\title{Plot Risk Profile (Type I and Type II Errors)}
\usage{
plot_risk(oc_data, x_var, facet_formula = NULL, ...)
}
\arguments{
\item{oc_data}{A data frame, typically from \code{run_simulation()}.}

\item{x_var}{A character string for the x-axis variable.}

\item{facet_formula}{A formula for faceting the plot (e.g., \code{~ trt_n}).}

\item{...}{Additional arguments passed to \code{ggplot2} functions.}
}
\value{
A ggplot object.
}
\description{
Creates a plot of the False Go Rate (FGR) and False Stop Rate (FSR).
}
