% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscore_func.R
\name{sd_gscore_median}
\alias{sd_gscore_median}
\title{Variance Estimator for g-score Median Estimate}
\usage{
sd_gscore_median(x)
}
\arguments{
\item{x}{A vector of g-scores.}
}
\value{
A scalar. Variance estimator for the g-score median.
}
\description{
This function calculates the variance estimator for the median estimate of g-scores, based on the method proposed by Price & Bonett (2001).
}
\examples{
g_scores <- simulate_gscore(n = 100, mu = 1, sigma = 0.5, prob_zero = 0.2)
sd_gscore_median(g_scores)
}
