% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscore_func.R
\name{simulate_gscore}
\alias{simulate_gscore}
\title{Simulate g-scores with Zero Inflation}
\usage{
simulate_gscore(n, mu, sigma, prob_zero)
}
\arguments{
\item{n}{A scalar. Number of patients in the arm of interest.}

\item{mu}{A scalar. Mean value of the normal distribution for generating the log of positive g-scores.}

\item{sigma}{A scalar. Standard deviation of the normal distribution for generating the log of positive g-scores.}

\item{prob_zero}{A scalar. Probability of a g-score being zero.}
}
\value{
A vector. Simulated g-scores.
}
\description{
This function generates g-scores by simulating positive values from a log-normal distribution, with added zero inflation.
The probability of a g-score being zero is controlled by \code{prob_zero}, while positive values are generated as log-normal.
}
\details{
P(G=0) = p, P(G>0) = 1-p;
log(G) | G > 0 ~ N(mu, sigma^2)
}
\examples{
simulate_gscore(n = 100, mu = 1, sigma = 0.5, prob_zero = 0.2)
}
