% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{simulate_ptss}
\alias{simulate_ptss}
\title{Simulate PTSS values from a transformed Beta distribution}
\usage{
simulate_ptss(n, mu, sigma, a = -100, b = 100)
}
\arguments{
\item{n}{Integer. Number of observations to simulate.}

\item{mu}{Numeric. Target mean PTSS on the [-100, 100] scale.}

\item{sigma}{Numeric. Target standard deviation PTSS on the [-100, 100] scale.}

\item{a}{Numeric. Lower bound of PTSS scale (default = -100).}

\item{b}{Numeric. Upper bound of PTSS scale (default = 100).}
}
\value{
A numeric vector of simulated PTSS values.
}
\description{
Simulate PTSS values from a transformed Beta distribution
}
\examples{
set.seed(123)
simulate_ptss_transformed_beta(n = 100, mu = 20, sigma = 20)
}
