% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_lalonde_decision.R
\name{bayesian_lalonde_decision}
\alias{bayesian_lalonde_decision}
\title{Bayesian Lalonde Decision Framework}
\usage{
bayesian_lalonde_decision(
  endpoint,
  data_summary,
  prior_params,
  lrv = 0,
  tv = 0,
  fgr = 0.2,
  fsr = 0.1,
  arm_names,
  posterior_infer = TRUE,
  Lalonde_decision = TRUE,
  EXP_TRANSFORM = FALSE
)
}
\arguments{
\item{endpoint}{A string. Type of endpoint ("continuous" or "binary").}

\item{data_summary}{A data frame. Summary of the individual-level data. For continuous endpoints, each row represents the number of patients, the mean of the normal distribution, and the standard deviation of the normal distribution for one arm in one simulation repetition. For binary endpoints, each row represents the number of patients and the number of responses for one arm in one repetition of the simulation.}

\item{prior_params}{A list. Prior distribution parameters for both the treatment and control arms. It should include delta, w, a, and b values for both arms.}

\item{lrv}{A scalar. The lower reference value for decision-making.}

\item{tv}{A scalar. The target value for decision-making.}

\item{fgr}{A scalar. False go rate for decision-making.}

\item{fsr}{A scalar. False stop rate for decision-making.}

\item{arm_names}{A list. Names of the arms in the simulation (e.g., treatment, control, etc.).}

\item{posterior_infer}{A logical. If TRUE, the function computes posterior inference for credible intervals and posterior probabilities.}

\item{Lalonde_decision}{A logical. If TRUE, the function makes decisions based on the Lalonde framework.}

\item{EXP_TRANSFORM}{A logical. If TRUE, the function exponentiates the results, which is useful for log-transformed data.}
}
\value{
A list. The output includes:
  - `post_params`: A data frame containing the posterior distribution parameters for each arm.
  - `post_est_ci`: A data frame containing evaluating metrics for the posterior distribution.
  - `post_inference`: A data frame containing the posterior probabilities and credible interval, and the decisions based on the Lalonde framework.
}
\description{
This function implements a Bayesian decision-making framework based on the Lalonde criteria.
It calculates posterior distributions for each arm and evaluates posterior inference and decision criteria based on the given endpoint type (continuous or binary).
}
