% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_post_dist_metrics.R
\name{calc_post_dist_metrics}
\alias{calc_post_dist_metrics}
\title{Calculate Posterior Distribution Metrics}
\usage{
calc_post_dist_metrics(endpoint, true_value, post_est_ci, remove.na = FALSE)
}
\arguments{
\item{endpoint}{A string. Specifies the type of endpoint ("g-score" or "OR").}

\item{true_value}{A scalar. The true median ratio for the "g-score" endpoint or the true rate difference for the "OR" endpoint.}

\item{post_est_ci}{A data frame. Contains posterior inference obtained from `bayesian_lalonde_decision()`, including posterior estimates, standard errors, and 95\% credible intervals for each simulation repetition.}

\item{remove.na}{}
}
\value{
A data frame containing the evaluation metrics for the posterior distribution, including:
  - `bias_avg`: The average bias of the estimate.
  - `sd_avg`: The average standard error of the estimate.
  - `sd_empirical`: The empirical standard deviation of the estimate.
  - `cp`: The nominal coverage probability (95\%).
  - Additional metrics specific to the "g-score" endpoint. More details can be found in `eval_gscore_approx_dist()`.
}
\description{
This function calculates various evaluation metrics for the posterior distribution, including the average bias, standard error, empirical standard deviation, and nominal coverage probability.
For the "g-score" endpoint, it calculates the log-median ratio, and for the "OR" endpoint, it calculates the rate difference.
}
