% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert_RBesT_mix}
\alias{convert_RBesT_mix}
\title{Convert to RBesT Mixture Distribution}
\usage{
convert_RBesT_mix(post, endpoint)
}
\arguments{
\item{post}{A data frame. Parameters for the mixture distribution, including the weight (`w`) of each component, and the distribution parameters for each component:
- For continuous endpoints: `w`, `mu1`, `sigma1`, `mu2`, and `sigma2`.
- For binary endpoints: `w`, `a1`, `b1`, `a2`, and `b2`.}

\item{endpoint}{A string. Specifies the type of endpoint ("continuous" for normal distributions, "binary" for beta distributions).}
}
\value{
An RBesT mixture distribution object. The object will be of class `mixnorm` for continuous endpoints or `mixbeta` for binary endpoints.
}
\description{
This function generates a mixture distribution in the format required by the RBesT package from a given set of parameters.
It supports both continuous and binary endpoints, constructing either normal or beta mixture distributions based on the endpoint type.
}
\examples{
# Example for continuous endpoint
post <- data.frame(w = 0.5, mu1 = 0, sigma1 = 1, mu2 = 0.5, sigma2 = 1)
mix_dist <- convert_RBesT_mix(post, endpoint = "continuous")

# Example for binary endpoint
post <- data.frame(w = 0.5, a1 = 1, b1 = 1, a2 = 2, b2 = 2)
mix_dist <- convert_RBesT_mix(post, "binary")
}
