% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_data_gen_params}
\alias{create_data_gen_params}
\title{Create Data Generation Parameters}
\usage{
create_data_gen_params(params, endpoint)
}
\arguments{
\item{params}{A list. Parameters for data generation.
- `trt_n`: A scalar. Sample size for the treatment group.
- `ctrl_n`: A scalar. Sample size for the control group. If not provided, it will be set to match `trt_n`.
- `trt_prob`: A scalar. Probability for the treatment group, interpretation based on the `endpoint`.
- `ctrl_prob`: A scalar. Probability for the control group, interpretation based on the `endpoint`.
- `trt_mu`: A scalar. Mean value for the treatment group (for non-zero values in g-scores or other continuous endpoints).
- `ctrl_mu`: A scalar. Mean value for the control group.
- `trt_sigma`: A scalar. Standard deviation for the treatment group (for non-zero values in g-scores or other continuous endpoints).
- `ctrl_sigma`: A scalar. Standard deviation for the control group.
- Additional parameters for historical data (e.g., `trt_h_n`, `ctrl_h_prob`) may also be included.}

\item{endpoint}{A string. Specifies the type of endpoint ("g-score" or "OR").}
}
\value{
A nested list. Processed parameters for each group, including `treatment`, `control`, `treatment_h`, and `control_h` if historical data is provided.
Missing values are replaced with `NA`, and a warning is issued for each missing parameter.
}
\description{
This function processes and validates a set of parameters for generating data
related to treatment and control groups, along with their historical counterparts.
It checks for missing values and replaces them with `NA` while issuing warnings.
The function supports different endpoints, and the interpretation of the `prob` parameter
varies depending on the specified endpoint:
  - For the "g-score" endpoint, `prob` refers to the probability of zero g-scores.
  - For the "OR" endpoint, `prob` refers to the probability of an event occurring.
}
\examples{
params <- list(trt_n = 100, trt_prob = 0.6, trt_mu = 1.5, trt_sigma = 0.3,
               ctrl_n = 100, ctrl_prob = 0.4, ctrl_mu = 1.2, ctrl_sigma = 0.2)
create_data_gen_params(params, "g-score")
}
