% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_all_functions}
\alias{get_all_functions}
\title{Retrieve All Functions from an Environment}
\usage{
get_all_functions(env = environment())
}
\arguments{
\item{env}{An environment. The environment to search for functions. The default is the current environment.}
}
\value{
A character vector. Names of all functions found in the specified environment.
}
\description{
This function returns the names of all functions within a specified environment.
}
