% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_OC.R
\name{obtain_OC}
\alias{obtain_OC}
\title{Obtain Operating Characteristics (OC)}
\usage{
obtain_OC(settings, decisions)
}
\arguments{
\item{settings}{A data frame. Contains the settings or parameters used in the simulations or trials.}

\item{decisions}{A data frame. Contains the decisions made in each simulation or trial, with columns for posterior probability-based decisions (`decision_pr`) and credible interval-based decisions (`decision_ci`).}
}
\value{
A data frame that includes the original settings along with the calculated proportions of decisions based on the posterior probability and credible interval. The output includes:
  - `proportion_pr`: Proportion of decisions made based on the posterior probability.
  - `proportion_ci`: Proportion of decisions made based on the credible interval.
}
\description{
This function calculates the operating characteristics (OC) by computing the proportion of decisions made based on the posterior probability and the credible interval from a set of decision results.
It combines these proportions with the provided settings data to return a comprehensive summary.
}
\examples{
# Example usage:
settings <- data.frame(setting1 = "A", setting2 = "B")
decisions <- data.frame(decision_pr = c("go", "no-go", "consider"),
                        decision_ci = c("go", "no-go", "consider"))
obtain_OC(settings, decisions)

}
