% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_oc}
\alias{plot_oc}
\title{Plot Operating Characteristics (OC)}
\usage{
plot_oc(
  oc_data,
  x_var,
  facet_formula = NULL,
  colors = c(`No-Go` = "red", Consider = "#F0E442", Go = "#009E73")
)
}
\arguments{
\item{oc_data}{A data frame, typically from \code{run_simulation()}.}

\item{x_var}{A character string for the x-axis variable.}

\item{facet_formula}{A formula for faceting the plot (e.g., \code{~ trt_n + true_value.compare_true}).}

\item{...}{Additional arguments passed to \code{ggplot2} functions.}
}
\value{
A ggplot object.
}
\description{
Creates a stacked bar plot of the operating characteristics.
}
