% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_distribution.R
\name{posterior_distribution}
\alias{posterior_distribution}
\title{Obtain Posterior Distribution for Continuous or Binary Endpoints}
\usage{
posterior_distribution(
  endpoint,
  current,
  historical = NULL,
  delta = log(0.85),
  w = NULL,
  a = 0.01,
  b = 0.01,
  a0 = 0.01,
  b0 = 0.01,
  theta0 = 0,
  s0 = 100,
  ess_h = NULL
)
}
\arguments{
\item{endpoint}{A string. The type of endpoint, either "continuous" or "binary".}

\item{current}{A named vector. Current trial data:
- For a continuous endpoint, the vector should include:
  - `n`: Number of patients.
  - `mu_hat`: Mean of the normal sampling distribution.
  - `s`: Standard deviation of the normal sampling distribution.
- For a binary endpoint, the vector should include:
  - `n`: Number of patients.
  - `count`: Number of responses.}

\item{historical}{A named vector. Historical trial data in the same format as the `current` vector. If `NULL`, borrowing is not performed.}

\item{delta}{A scalar. The clinically significant difference for the SAM prior, used as the threshold for detecting prior-data conflict (default is `log(0.85)` for continuous endpoints).}

\item{w}{A scalar. Initial weight assigned to the informative component of the mixture prior. Set to `0` if no borrowing. If `NULL`, the SAM prior is used for automatic conflict detection.}

\item{a}{A scalar. The alpha parameter of the beta prior for binary endpoints (used in the mixture prior if borrowing).}

\item{b}{A scalar. The beta parameter of the beta prior for binary endpoints (used in the mixture prior if borrowing).}

\item{gate}{A scalar.}
}
\value{
A list containing the posterior distribution parameters:
  - For continuous endpoints:
    - `w_prior`: Weight of the prior after conflict detection (or user-defined weight).
    - `post`: A list with:
      - `w`: Posterior weight of the informative component.
      - `theta1`, `sigma1`: Posterior mean and standard deviation from informative prior.
      - `theta2`, `sigma2`: Posterior mean and standard deviation from non-informative prior.
  - For binary endpoints:
    - `w_prior`: Weight of the prior after conflict detection (or user-defined weight).
    - `post`: A list with:
      - `w`: Posterior weight of the informative component.
      - `a1`, `b1`: Parameters of the informative beta posterior.
      - `a2`, `b2`: Parameters of the non-informative beta posterior.
}
\description{
This function calculates the posterior distribution for one-arm trial data with either continuous or binary endpoints, using current and optionally historical data.
It supports the SAM prior and allows for optional borrowing from historical data.
}
