% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_inference.R
\name{posterior_inference}
\alias{posterior_inference}
\title{Perform Posterior Inference}
\usage{
posterior_inference(post1, post2 = NULL, quantiles, EXP_TRANSFORM = FALSE)
}
\arguments{
\item{post1}{A RBesT mixture distribution object. The first posterior distribution (e.g., for the treatment arm).}

\item{post2}{A RBesT mixture distribution object. The second posterior distribution (e.g., for the control arm). Default is `NULL`, in which case only posterior estimates for `post1` are calculated.}

\item{quantiles}{A vector. The upper and lower quantiles for the credible interval (e.g., `c(0.025, 0.975)` for a 95\% credible interval).}

\item{EXP_TRANSFORM}{A logical. If TRUE, the function exponentiates the results, which is useful for log-transformed data.}
}
\value{
A data frame with one row containing:
  - \code{est1}: Posterior mean for the first posterior distribution (e.g., treatment arm).
  - \code{sd1}: Posterior standard deviation for the first posterior distribution.
  - \code{ci1_l}: Lower bound of the credible interval for the first posterior distribution.
  - \code{ci1_u}: Upper bound of the credible interval for the first posterior distribution.
  - \code{est2}: Posterior mean for the second posterior distribution (if provided).
  - \code{sd2}: Posterior standard deviation for the second posterior distribution (if provided).
  - \code{ci2_l}: Lower bound of the credible interval for the second posterior distribution (if provided).
  - \code{ci2_u}: Upper bound of the credible interval for the second posterior distribution (if provided).
  - \code{est_compare}: Estimated treatment effect, either as a difference (for continuous data) or a ratio (for log-transformed data if `EXP_TRANSFORM = TRUE`).
  - \code{sd_compare}: Standard deviation of the estimated treatment effect.
  - \code{compare_ci_l}: Lower bound of the credible interval for the treatment effect.
  - \code{compare_ci_u}: Upper bound of the credible interval for the treatment effect.
}
\description{
This function calculates posterior estimates, standard deviations, and credible intervals for individual arms and the treatment effect, based on posterior distributions from the RBesT package.
If a second posterior distribution is not provided, it calculates these statistics for a single arm.
The function also supports exponentiating the results, which can be useful for endpoints like log-transformed data.
}
