% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_prob.R
\name{posterior_prob}
\alias{posterior_prob}
\title{Calculate Posterior Probability}
\usage{
posterior_prob(
  post1,
  value,
  post2 = NULL,
  range_type = c("greater", "less", "between"),
  value2 = NULL
)
}
\arguments{
\item{post1}{A RBesT mixture distribution object. Represents the first posterior distribution.}

\item{value}{A scalar. The threshold or lower bound for the comparison.}

\item{post2}{A RBesT mixture distribution object. Optional second posterior distribution object for pairwise comparison. Default is `NULL`.}

\item{range_type}{A string. Specifies the type of comparison:
- "greater": Computes the probability that the posterior or the difference between `post1` and `post2` exceeds `value`.
- "less": Computes the probability that the posterior or the difference between `post1` and `post2` is less than `value`.
- "between": Computes the probability that the posterior or the difference between `post1` and `post2` lies between `value` and `value2`.}

\item{value2}{A scalar. The upper bound for the "between" comparison. Required if `range_type` is "between".}
}
\value{
A scalar. Posterior probability of the point estimate or the difference between estimates falling within the specified range.
}
\description{
This function calculates the posterior probability of a point estimate falling within a specified range.
If two posterior distributions are provided, it computes the posterior probability of the difference
between the two distributions falling within the range.
For a single posterior distribution, it computes the probability of the estimate exceeding, being less than, or falling between specified values.
}
\examples{
# Example: Posterior probability for a single distribution
post1 <- convert_RBesT_mix(post = data.frame(w = 0.5, mu1 = 0,
                                             sigma1 = 1, mu2 = 0.5, sigma2 = 1),
                           endpoint = "continuous")
posterior_prob(post1, value = 0.5, range_type = "greater")

# Example: Posterior probability for the difference between two distributions
post2 <- convert_RBesT_mix(post = data.frame(w = 0.9, mu1 = 0, sigma1 = 1,
                                             mu2 = 0.5, sigma2 = 1),
                           endpoint = "continuous")
posterior_prob(post1, value = 0.5, post2 = post2, range_type = "between", value2 = 1)

}
